+++ .gitignore
@@ -2,4 +2,4 @@
 /node_modules
 .idea/
 yarn.lock
-utils.js
+object.js
