+++ .npmignore
@@ -2,4 +2,4 @@
 /node_modules
 .idea/
 yarn.lock
-utils.ts
+object.ts
